Add discover page views
This is a migration to the proper pattern for react component management.

Barebone/pure components are in `src/components`. These are "composed" to build a view, in `src/views`, and are finally used by `src/pages`. Pretty standard pattern, and now is a good time to bring it in.

React Components -> Composed Views -> Rendered Page.