Start covid specific API router
I am trying a new approach in designing proper
module specs. Idea is to 1: keep such experiments least impactful on
prod systems, and 2: allowing for using all that we've developed
together.

A quick explanation here is:

For this module, I need to have a particular server behaviour, and
corresponding UI components to behave properly. I also want to
reuse as many old code as applicable. But not at the expense of
either breaking working pages or compromising the quality of the
UI I am working on, right now. So, instead, I decided to work on
a copy of the old module*, and if this experiment is succesful,
I should be able to define/refine the specs for the code as whole.
We will bring the changes back in main branch with proper review.

[*] Notice I say module, not a code fragment. I am doing this to
avoid making copy of functions/components go anywhere near
prod.